+++ jwst/srctype/srctype.py
@@ -8,7 +8,6 @@ log.setLevel(logging.DEBUG)
 
 
 def set_source_type(input_model, source_type=None):
-
     """
     Set source_type based on APT input, user specification, exposure type,
     or default values.
@@ -16,19 +15,17 @@ def set_source_type(input_model, source_type=None):
     Parameters
     ----------
     input_model : `~jwst.datamodels.CubeModel`, `~jwst.datamodels.ImageModel`,
-                  `~jwst.datamodels.IFUImageModel`,
-                  `~jwst.datamodels.MultiSlitModel`,
+                  `~jwst.datamodels.IFUImageModel`, `~jwst.datamodels.MultiSlitModel`,
                   or `~jwst.datamodels.SlitModel`
         The data model to be processed.
 
-    source_type : string ['POINT'|'EXTENDED']; optional
-        User-requested value for source type
+    source_type : str,  {POINT, EXTENDED}
+        User-requested value for source type.
 
     Returns
     -------
     input_model : `~jwst.datamodels.CubeModel`, `~jwst.datamodels.ImageModel`,
-                  `~jwst.datamodels.IFUImageModel`,
-                  `~jwst.datamodels.MultiSlitModel`,
+                  `~jwst.datamodels.IFUImageModel`, `~jwst.datamodels.MultiSlitModel`,
                   or `~jwst.datamodels.SlitModel`
         The updated model.
     """
@@ -40,12 +37,11 @@ def set_source_type(input_model, source_type=None):
         raise RuntimeError('Step cannot be executed without an EXP_TYPE value')
     else:
         log.info(f'Input EXP_TYPE is {exptype}')
-
     # For exposure types that have a single source specification, get the
     # user-supplied source type from the selection they provided in the APT
     if exptype in ['MIR_LRS-FIXEDSLIT', 'MIR_LRS-SLITLESS', 'MIR_MRS',
-                   'NRC_TSGRISM', 'NIS_SOSS', 'NRS_FIXEDSLIT', 'NRS_BRIGHTOBJ',
-                   'NRS_IFU']:
+                   'NRC_TSGRISM', 'NIS_SOSS', 'NRS_FIXEDSLIT',
+                   'NRS_BRIGHTOBJ', 'NRS_IFU']:
 
         # Get info about the exposure, including whether it's a background
         # target and the dither pattern type
@@ -65,23 +61,34 @@ def set_source_type(input_model, source_type=None):
             user_type = input_model.meta.target.source_type_apt
             log.info(f'Input SRCTYAPT = {user_type}')
             if user_type is None:
-                log.warning('SRCTYAPT keyword not found in input; using \
-                             SRCTYPE instead')
+                log.warning('SRCTYAPT keyword not found in input; using SRCTYPE instead')
                 user_type = input_model.meta.target.source_type
                 input_model.meta.target.source_type_apt = user_type
         except AttributeError:
-            log.warning('SRCTYAPT keyword not found in input; using SRCTYPE \
-                        instead')
+            log.warning('SRCTYAPT keyword not found in input; using SRCTYPE instead')
             user_type = input_model.meta.target.source_type
             input_model.meta.target.source_type_apt = user_type
 
-        if bkg_target:
+        # Check to see if the user specified a source type
+        if source_type is not None:
+            source_type = str(source_type).upper()
+
+            # Check if the exposure type is a mode that allows setting
+            if exptype in ['MIR_LRS-FIXEDSLIT', 'MIR_LRS-SLITLESS',
+                           'MIR_MRS', 'NRC_TSGRISM', 'NRS_FIXEDSLIT',
+                           'NRS_BRIGHTOBJ', 'NRS_IFU']:
+
+                src_type = source_type
+
+            log.warning(f'Based on user-input, setting SRCTYPE = {src_type}')
+            input_model.meta.target.source_type = src_type
+
+        elif bkg_target:
 
             # If this image is flagged as a BACKGROUND target, set the
             # source type to EXTENDED regardless of any other settings
             src_type = 'EXTENDED'
-            log.info(f'Exposure is a background target; setting SRCTYPE \
-                       = {src_type}')
+            log.info(f'Exposure is a background target; setting SRCTYPE = {src_type}')
 
         elif pipe_utils.is_tso(input_model):
 
@@ -89,8 +96,7 @@ def set_source_type(input_model, source_type=None):
             src_type = 'POINT'
             log.info(f'Input is a TSO exposure; setting SRCTYPE = {src_type}')
 
-        elif (patttype is not None) and (('NOD' in patttype) or
-                                         ('POINT-SOURCE' in patttype)):
+        elif (patttype is not None) and (('NOD' in patttype) or ('POINT-SOURCE' in patttype)):
 
             # Set all nodded exposures to POINT source type
             src_type = 'POINT'
@@ -110,29 +116,7 @@ def set_source_type(input_model, source_type=None):
             else:
                 src_type = 'POINT'
 
-            log.info(f'Input source type is unknown; setting default SRCTYPE \
-                     = {src_type}')
-
-        # Check to see if the user specified a source type
-        if source_type is not None:
-            source_type = str(source_type).upper()
-
-            # Check if requested type is one of the 2 allowed types
-            if source_type != 'POINT' and source_type != 'EXTENDED':
-                log.warning(f'Requested source type {source_type} not \
-                            applicable to this operation')
-
-                src_type = 'UNKNOWN'
-
-            # Check if the exposure type is a mode that allows setting
-            elif exptype in ['MIR_LRS-FIXEDSLIT', 'MIR_LRS-SLITLESS',
-                             'MIR_MRS', 'NRC_TSGRISM', 'NRS_FIXEDSLIT',
-                             'NRS_BRIGHTOBJ', 'NRS_IFU']:
-
-                src_type = source_type
-
-            log.warning(f'Based on user-input, setting SRCTYPE = {src_type}')
-            input_model.meta.target.source_type = src_type
+            log.info(f'Input source type is unknown; setting default SRCTYPE = {src_type}')
 
         # Set the source type in the global meta attribute
         input_model.meta.target.source_type = src_type
@@ -177,8 +161,7 @@ def set_source_type(input_model, source_type=None):
             else:
                 slit.source_type = 'EXTENDED'
 
-            log.info(f'source_id={slit.source_id}, \
-                      stellarity={stellarity:.4f}, type={slit.source_type}')
+            log.info(f'source_id={slit.source_id}, stellarity={stellarity:.4f}, type={slit.source_type}')
 
         # Remove the global target source type, so that it never mistakenly
         # gets used for MOS data, which should always use slit-specific values
@@ -187,8 +170,7 @@ def set_source_type(input_model, source_type=None):
     # Set all TSO exposures to POINT
     elif pipe_utils.is_tso(input_model):
         src_type = 'POINT'
-        log.info(f'Input is a TSO exposure; setting default SRCTYPE = \
-                  {src_type}')
+        log.info(f'Input is a TSO exposure; setting default SRCTYPE = {src_type}')
         input_model.meta.target.source_type = src_type
 
     # For WFSS modes check slit values of is_extended to set SRCTYPE